requirements.txt
PySide6==6.7.2
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
openpyxl==3.1.5
bcrypt==4.2.0
appdirs==1.4.4
Pillow==11.0.0
